# Optimize GLES3 CanvasLight calculations and make it precise.