@jss2a98aj 2024-11-30
Can you upload a project with that test scene? I would like to benchmark it.