@WhalesState 2024-11-30
I have edited the project to use the default physics engine, and freezed all the physics nodes after 0.5 seconds because godot 2d physics performance will affect the benchmark.

[test.zip](https://github.com/user-attachments/files/17964886/test.zip)