@jss2a98aj 2024-11-30
I am getting slightly worse performance in the test with the changes. I am going to try production builds next to see if it changes anything.

Before
```
Project FPS: 1216 (0.82 mspf)
Project FPS: 1249 (0.80 mspf)
Project FPS: 1250 (0.80 mspf)
Project FPS: 1244 (0.80 mspf)
Project FPS: 1215 (0.82 mspf)
Project FPS: 1210 (0.82 mspf)
Project FPS: 1238 (0.80 mspf)
```

After
```
Project FPS: 1178 (0.84 mspf)
Project FPS: 1178 (0.84 mspf)
Project FPS: 1170 (0.85 mspf)
Project FPS: 1169 (0.85 mspf)
Project FPS: 1179 (0.84 mspf)
Project FPS: 1156 (0.86 mspf)
Project FPS: 1145 (0.87 mspf)
```