@WhalesState 2024-12-01
I have done some more changes to make it slightly faster.
Building both locally with `optimize=none`.

// Before
Project FPS: 103 (9.70 mspf)
Project FPS: 102 (9.80 mspf)
Project FPS: 105 (9.52 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 105 (9.52 mspf)
Project FPS: 108 (9.25 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 108 (9.25 mspf)

// After
Project FPS: 113 (8.84 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 108 (9.25 mspf)
Project FPS: 104 (9.61 mspf)
Project FPS: 110 (9.09 mspf)
Project FPS: 112 (8.92 mspf)
Project FPS: 105 (9.52 mspf)
Project FPS: 113 (8.84 mspf)